% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hgFemale}
\alias{hgFemale}
\title{Human female tissue polygon data}
\format{
See \link{hgMale}.
}
\source{
\url{https://github.com/ebi-gene-expression-group/anatomogram}
(images CC-BY-4.0, code Apache-2.0). See \code{NOTES.md} for attribution
and parsing details.
}
\usage{
hgFemale
}
\description{
Same structure as \link{hgMale}, for the human female anatomogram.
}
\keyword{datasets}
